// src/app/module/coverage/enums/fields.enu.ts
export enum FieldNames {
  ORIGIN = 'origin',
  PRODUCT = 'product',
  FOLIO_NUMBER = 'folio_number',
  OPERATOR = 'operator',
  INSPECTION_LOT = 'inspection_lot',
  HOPPER_COMPARTMENT = 'compartimientotolva',
  FREIGHT_CAR_COMPARTMENT = 'compartimientofurgon',
  CARRIER = 'carrier',
  PACKAGE_CODE = 'package_code',
  VEHICLE_CODE = 'vehicle_code',
  SAMPLE_TYPE = 'sample_type',
  LINE = 'line',
  ANALYSIS = 'analysis',
  AREA = 'area',
}

export enum FieldTypes {
  SELECTOR = 'selector',
  TEXT = 'text',
  VALIDATOR = 'validator',
}